+++ src/main/java/com/example/capstone/repository/Product/IReviewRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IReviewRepository extends JpaRepository<Favorite, Integer> {
+public interface IReviewRepository extends JpaRepository<Favorite, Long> {
 
 }
