+++ src/main/java/com/example/capstone/repository/Veterinarian/IPet_IssueRepository.java
@@ -4,5 +4,5 @@ package com.example.capstone.repository.Veterinarian;
 import com.example.capstone.model.Veterinarian.PetIssue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPet_IssueRepository extends JpaRepository<PetIssue, Integer> {
+public interface IPet_IssueRepository extends JpaRepository<PetIssue, Long> {
 }
